remove `frequency` parameter from `enableBackgroundDelivery` function
why? we always were passing in .immediate anyway (it's the default value and we never actually specify the param at a call site), and since we might have multiple `enableBackgroundDelivery` calls ending up with the same HKHealthStore-internal background delivery thing, we're already not handling this properly (imagine calling the function 2 times for the same sample types, but different delivery frequencies. only one of the 2 frequencies would get applied)